+++ src/main/java/com/sparta/gathering/common/config/aws/AwsConfigure.java
@@ -1,4 +1,4 @@
-package com.sparta.gathering.common.config;
+package com.sparta.gathering.common.config.aws;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
