+++ src/main/java/com/sparta/gathering/common/config/aws/S3Client.java
@@ -1,4 +1,4 @@
-package com.sparta.gathering.common.config;
+package com.sparta.gathering.common.config.aws;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
