+++ src/main/resources/database/legacy/schema_v0.sql
